Se Configuraron y se corrijieron los roles en jwt, el usuario se loguea con su correo y password, mientras que el admin con su cc y password, se organizaron y se crearon nuevas clases, se esta llevando a cabo la relacion de tablas con jpa de uno a muchos y muchos a uno, de igual forma se crearon nueva tablas en la bd

+++ src/main/java/com/example/parking_management/controller/FeeController.java
@@ -28,7 +28,7 @@ public class FeeController {
 
 
     @GetMapping("/{feeId}")
-    public Optional<Fee> getById(@PathVariable("feeId") Long feeId)
+    public Optional<Fee> getById(@PathVariable("feeId") int feeId)
     {
         return feeService.getFee(feeId);
     }
@@ -42,7 +42,7 @@ public class FeeController {
 
 
     @DeleteMapping("/{feeId}")
-    public void saveOrUpdate(@PathVariable("freeId") Long feeId)
+    public void saveOrUpdate(@PathVariable("freeId") int feeId)
     {
         feeService.delete(feeId);
     }
