+++ src/main/java/com/example/parking_management/repository/FeeRepository.java
@@ -6,7 +6,7 @@ import org.springframework.stereotype.Repository;
 
 
 @Repository
-public interface FeeRepository extends JpaRepository<Fee, Long>
+public interface FeeRepository extends JpaRepository<Fee, Integer>
 {
 
 }
